+++ packages/ui-kit-js/src/types.d.ts
@@ -1,37 +1,5 @@
-// global text variants
-export type TextVariant =
-  | 'copy'
-  | 'small'
-  | '18'
-  | '20'
-  | '20-thin'
-  | '24'
-  | '24-thin'
-  | '28'
-  | '28-thin'
-  | '30'
-  | '30-thin'
-  | '32'
-  | '32-thin'
-  | '36'
-  | '36-thin'
-  | '42'
-  | '42-thin'
-  | '44'
-  | '44-thin'
-  | '48'
-  | '48-thin'
-  | '52'
-  | '52-thin'
-  | '60'
-  | '60-thin'
-  | '62'
-  | '62-thin'
-  | '72'
-  | '72-thin'
-  | '84'
-  | '84-thin';
+// global text sizes
+export type TextSize = 'x-small' | 'small' | 'medium' | 'large' | 'x-large' | 'inherit';
 
 // global text colors
-export type TextColor = 'inherit' | 'porsche-black' | 'porsche-light' | 'porsche-red';
-
+export type TextColor = 'porsche-black' | 'porsche-light' | 'porsche-red' | 'inherit';